requirements.txt
beautifulsoup4>=4.6.0
PyYAML>=3.12
requests>=2.18.4
GitPython>=2.1.7
